Rename lidspace compaction job
